Update files/directories to reduce lint-check exclusions (iiasa/message_data#384)
* apply code style to
  * message_data/scenario_generation
  * message_data/tools/post_processing
  * message_data/tools/prep_submission.py
  * message_data/tools/utilities
  * message_data/model/cli and create
  * message_data/projects/cd_links
  * message_data/projects/engage
  * message_data/tests/tools
* apply isort to tests
* update pyproject.toml to remove message_data/model/cli and create
* remove isort exemptions for all tests
* remove isort exemptions for message_data/tools/*.py

+++ message_ix_models/tools/remove_emission_bounds.py
@@ -1,5 +1,5 @@
 def main(scen, remove_all=False):
-    """ Removes parameters 'tax_emission' and 'bound_emission'
+    """Removes parameters 'tax_emission' and 'bound_emission'
     from a given scenario.
 
     Parameters
@@ -9,23 +9,26 @@ def main(scen, remove_all=False):
     """
 
     scen.check_out()
-    for par in ['bound_emission', 'tax_emission']:
+    for par in ["bound_emission", "tax_emission"]:
         df = scen.par(par)
         if not df.empty:
             # Remove cumulative years
-            df_cum = df[df.type_year == 'cumulative']
+            df_cum = df[df.type_year == "cumulative"]
             if not df_cum.empty:
                 scen.remove_par(par, df_cum)
-                df = df[df.type_year != 'cumulative']
+                df = df[df.type_year != "cumulative"]
             # Remove yearly bounds
             if not remove_all:
-                df['type_year'] = df['type_year'].astype('int64')
-                df = df[df['type_year'] > int(
-                    scen.set('cat_year',
-                             {'type_year': ['firstmodelyear']})['year'])]
+                df["type_year"] = df["type_year"].astype("int64")
+                df = df[
+                    df["type_year"]
+                    > int(
+                        scen.set("cat_year", {"type_year": ["firstmodelyear"]})["year"]
+                    )
+                ]
             scen.remove_par(par, df)
-    scen.commit('removed emission bounds and taxes')
+    scen.commit("removed emission bounds and taxes")
 
 
-if __name__ == '__main__':
-    main('test', 'test')
+if __name__ == "__main__":
+    main("test", "test")
